LPS-196211 Move ObjectDefinitionUtil to object-api
